company_pitchbook/requirements.txt
streamlit>=1.40.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.2.1
pydantic>=2.6.4
reportlab>=4.1.0
brotli>=1.1.0

# Data visualization
matplotlib>=3.8.2
seaborn>=0.13.2
pillow>=10.2.0

# LangChain ecosystem
langchain==0.3.10
langchain-openai==0.2.11
langchain-community==0.3.10
langchain-core==0.3.22
langgraph==0.2.56
langsmith>=0.1.8

# Search and extraction
duckduckgo-search>=4.4.3
unstructured>=0.11.0
pypdf>=4.0.0
lxml>=5.1.0
html5lib>=1.1
faiss-cpu>=1.7.4

# Async and networking
aiohttp>=3.8.0,<4.0.0
aiofiles>=23.2.1

# Type stubs for development
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0
types-aiofiles>=23.2.0
types-setuptools>=69.0.0
types-html5lib>=1.1.11
types-lxml>=2024.2.9
types-reportlab>=3.6.17
types-urllib3>=1.26.25
types-python-dateutil>=2.8.19
types-Pillow>=10.0.0
types-PyYAML>=6.0.12

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Tokenization
tiktoken>=0.5.2

# Development tools
mypy>=1.8.0
black>=24.2.0
isort>=5.13.0
